did loginpage modular
